refactor: extract helpers and cellmap class
