+++ src/crossworddom.js
@@ -1,64 +1,5 @@
-//  Lightweight helper functions.
-function removeClass(element, className) {
-  const expression = new RegExp(`(?:^|\\s)${className}(?!\\S)`, 'g');
-  element.className = element.className.replace(expression, '');
-}
-function addClass(element, className) {
-  element.className += ` ${className}`;
-}
-
-//  Internally used map of Crossword model data to DOM elements.
-function CellMap() {
-  this.map = [];
-}
-
-//  Adds a Cell <-> Cell Element mapping.
-CellMap.prototype.add = function (cell, cellElement) {
-  this.map.push({
-    cell,
-    cellElement,
-  });
-};
-
-//  Gets the DOM element for a cell.
-CellMap.prototype.getCellElement = function (cell) {
-  for (let i = 0; i < this.map.length; i++) {
-    if (this.map[i].cell === cell) {
-      return this.map[i].cellElement;
-    }
-  }
-  return null;
-};
-
-//  Gets the cell for a DOM element.
-CellMap.prototype.getCell = function (cellElement) {
-  for (let i = 0; i < this.map.length; i++) {
-    if (this.map[i].cellElement === cellElement) {
-      return this.map[i].cell;
-    }
-  }
-  return null;
-};
-
-
-//  Removes entries for a crossword.
-CellMap.prototype.removeCrosswordCells = function removeCrosswordCells(crossword) {
-  for (let i = 0; i < this.map.length; i++) {
-    if (this.map[i].cell.crossword === crossword) {
-      this.map.splice(i, 1);
-    }
-  }
-};
-
-const getCellElementData = function getCellElementData(cellElement) {
-  for (let i = 0; i < cellMap.length; i++) {
-    if (cellMap[i].cellElement === cellElement) {
-      return cellMap[i];
-    }
-  }
-  return null;
-};
-
+const CellMap = require('./cell-map.js');
+const { removeClass, addClass } = require('./helpers');
 //  Creates the DOM representation of a Crossword.
 function CrosswordDOM(crossword, parentElement) {
   this.crossword = crossword;
@@ -382,6 +323,6 @@ CrosswordDOM.prototype._updateDOM = function _updateDOM() {
 };
 
 //  Create a single global instance of a cell map.
-var cellMap = new CellMap();
+const cellMap = new CellMap();
 
 module.exports = CrosswordDOM;
